+++ src/core/util/fst_impl/bytes_ref_fst_enum.rs
@@ -65,7 +65,7 @@ where
 
   pub fn seek_ceil(&mut self, target: &BytesRef<Vec<u8>>) -> Result<Option<&IOBytesRef<O>>> {
     let (base, result) = (&mut self.base, &mut self.result);
-    base.target_length = target.length as i32;
+    base.target_length = target.length;
     let mut ctx = BytesRefFSTEnumContext { result };
     base.do_seek_ceil(&mut ctx, target)?;
     self.set_result()
@@ -73,7 +73,7 @@ where
 
   pub fn seek_floor(&mut self, target: &BytesRef<Vec<u8>>) -> Result<Option<&IOBytesRef<O>>> {
     let (base, result) = (&mut self.base, &mut self.result);
-    base.target_length = target.length as i32;
+    base.target_length = target.length;
     let mut ctx = BytesRefFSTEnumContext { result };
     base.do_seek_floor(&mut ctx, target)?;
     self.set_result()
@@ -81,7 +81,7 @@ where
 
   pub fn seek_exact(&mut self, target: &BytesRef<Vec<u8>>) -> Result<Option<&IOBytesRef<O>>> {
     let (base, result) = (&mut self.base, &mut self.result);
-    base.target_length = target.length as i32;
+    base.target_length = target.length;
     let mut ctx = BytesRefFSTEnumContext { result };
     let found = base.do_seek_exact(&mut ctx, target)?;
     if found {
